Safari-609.1.11.1/LayoutTests/imported/blink/http/tests/css/remove-placeholder-styles.html

<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

var link = document.createElement('link');
link.setAttribute('rel', 'stylesheet');
link.setAttribute('href', 'resources/slow-loading-sheet.php?color=green');
document.head.appendChild(link);
alert(document.documentElement.offsetWidth);
</script>
This test passes if it alerts 800.
