third_party/blink/web_tests/platform/generic/http/tests/css/remove-placeholder-styles-expected.txt

ALERT: 800
This test passes if it alerts 800.
